+++ src/main/java/ru/dst/analyze/locarus/handlers/FanDrivePressHandler.java
@@ -47,7 +47,7 @@ public class FanDrivePressHandler extends Handler{
     }
 
     @Override
-    public int getMultiply() {
+    public double getMultiply() {
         return multiply;
     }
