+++ src/main/java/ru/dst/analyze/locarus/handlers/TempTurboHandler.java
@@ -47,7 +47,7 @@ public class TempTurboHandler extends Handler {
     }
 
     @Override
-    public int getMultiply() {
+    public double getMultiply() {
         return multiply;
     }
